+++ src/app/admin/honeymoon/checklist/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import ChecklistTab from '../ChecklistTab';
+import { useHoneymoonApi } from '../HoneymoonContext';
+
+export default function HoneymoonChecklistPage() {
+    return <ChecklistTab api={useHoneymoonApi()} />;
+}